requirements.txt
streamlit
chromadb
sentence-transformers
PyPDF2
pillow
python-dotenv
youtube-transcript-api
python-pptx
pytesseract
llama-index
crewai
crewai-tools
SpeechRecognition
PyAudio
gTTS
pydub
sounddevice
SpeechRecognition
webrtcvad
numpy
git+https://github.com/openai/whisper.git
pyttsx3
transformers 
torch